pyproject.toml
[tool.poetry]
name = "LightAutoML"
version = "0.2.16"
description = "Fast and customizable framework for automatic ML model creation (AutoML)"
authors = ["Alexander Ryzhkov <AMRyzhkov@sberbank.ru>",
           "Vakhrushev Anton <AGVakhrushev@sberbank.ru>",
           "Simakov Dmitrii <Simakov.D.E@sberbank.ru>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://lightautoml.readthedocs.io/en/latest/"
repository = "https://github.com/sberbank-ai-lab/LightAutoML"
classifiers = [
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Natural Language :: Russian",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed"
]


[tool.poetry.dependencies]
python = "^3.6.1"

poetry-core = "^1.0.0"
numpy = [
  {version = ">=1.20.0", python = ">=3.7"},
  {version = "*", python = "<3.7"}
]
scipy = "*"
pandas = ">=1"
scikit-learn = ">=0.22"
lightgbm = "^2.3"
optuna = "*"
torch = "<1.9"
dataclasses = "0.6"
holidays = "*"
networkx = "*"
cmaes = "*"
log-calls = "*"
pyyaml = "*"
catboost = "*"

autowoe = ">=1.2"

json2html = "*"
jinja2 = "*"
seaborn = "*"

transformers = ">=4"
nltk = "*"
gensim = ">=4"
tqdm = "*"
efficientnet-pytorch = "*"
albumentations = ">=0.4.6"
opencv-python = "*"
joblib = "*"
pywavelets = "*"
scikit-image = "*"
torchvision = "*"

webencodings = "*"

importlib-metadata = {version = "^1.0", python = "<3.8"}
weasyprint = {version = "^52.5", optional = true}
cffi = {version = "^1.14.5", optional = true}


[tool.poetry.dev-dependencies]
pytest = "*"
sphinx = "*"
sphinx-rtd-theme = "*"
IPython = "*"
nbsphinx = "*"
nbsphinx-link = "*"
sphinx-autodoc-typehints = "*"
pandoc = "*"
pre-commit = "*"


[tool.poetry.extras]
pdf = ["weasyprint", "cffi"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
filter_files = true


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
